+++ src/routes/comment/[comment_id]/+page.server.ts
@@ -5,7 +5,7 @@ import {
 	addCommentsLikes,
 	getCommentsLikesByCommentId,
 	updateCommentslikes
-} from '$lib/stores/comments.store';
+} from '$lib/db/comments.store';
 
 export const actions: Actions = {
 	like: async (e) => {
