+++ src/routes/signout/+page.server.ts
@@ -1,4 +1,4 @@
-import { deleteSession } from '$lib/stores/sessions.store';
+import { deleteSession } from '$lib/db/sessions';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = (e) => {
